#15 [REFACTOR] User 엔티티 위치 변경
